omap_tiler_heap: Fix sg list for dynamic tiler allocations
The scattergather list for dynamic tiler allocations were not
created properly. TI forgot to call sg_next, so only the first
item in the list were changed. This caused the table to look
like this:

    [0] = { length: 4096 }
    [1] = { length : 0 }
    [2] = { length : 0 }

and so on, instead of:

    [0] = { length: 4096 }
    [1] = { length: 4096 }
    [2] = { length: 4096 }

which is what we expected. This caused the SGX graphics driver to
overwrite memory, as the sg list is used to calculate the length
of an array that it uses and we didn't do range checking on that,
and that further caused things to crash if we started the camera
app.

Change-Id: I317754db87e19f42508d8b8f631c71ef0dbb805c
Signed-off-by: D. Andrei Măceș <andrei@unlegacy-android.org>